Support TTN ABP on US915
I have TTN working with this device using ABP and on the US915
frequencies. OTAA is untested. I cannot test EU868 or other bands.

All tx channels have been disabled aside from 8. Alternative SF
for US915 are used when selected to support "CR".

Only use custom RX2 window for EU868 for TTN when using ABP.
Works fine in US915 when this is disabled, unsure if it re-
activates for EU868.